Desktop/sistema-reparaciones/requirements.txt
# requirements.txt
# Dependencias para el Sistema de Gestión de Reparaciones

# Framework principal
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Migrate==4.0.5

# Seguridad y formularios
WTForms==3.0.1
Werkzeug==2.3.7

# Variables de entorno
python-dotenv==1.0.0

# Base de datos
SQLAlchemy==2.0.21

# Utilidades
email-validator==2.0.0

# Para producción
gunicorn==21.2.0